refactor: move jwtAuthFilter in the security package
