+++ backend/src/main/java/ch/sectioninformatique/template/security/JwtAuthFilter.java
@@ -1,4 +1,4 @@
-package ch.sectioninformatique.template.jwt;
+package ch.sectioninformatique.template.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
